refactor(cli): split app shell from core logic
Move reusable bibliography conversion and lint behavior into engine
and schema library APIs, leaving citum-cli focused on argument parsing,
command dispatch, terminal output, and style resolution.

Schema-Bump: patch

Version-Bump: minor
